roles, default statuses, permissions set

+++ eventMS/src/main/java/org/example/event/EventService.java
@@ -9,22 +9,16 @@ import org.example.event.dto.ProjectResponse;
 import org.example.event.dto.UserResponse;
 import org.example.event_type.EventType;
 import org.example.event_type.EventTypeRepository;
-import org.example.event_type.EventTypeService;
-import org.example.event_type.dto.EventTypeResponse;
 import org.example.mappers.EventMapper;
-import org.example.mappers.EventTypeMapper;
-import org.example.mappers.StatusMapper;
 import org.example.status.Status;
 import org.example.status.StatusRepository;
-import org.example.status.StatusService;
-import org.example.status.dto.StatusResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
 import org.springframework.stereotype.Service;
 
 import java.time.OffsetDateTime;
-import java.util.ArrayList;
 import java.util.HashSet;
+import java.util.Objects;
 import java.util.Set;
 
 @Service
@@ -39,6 +33,8 @@ public class EventService {
     EventTypeRepository eventTypeRepository;
     @Autowired
     StatusRepository statusRepository;
+
+    private final String DEFAULT_STATUS = "todo";
 
     public void addParticipant(Integer userId, Integer eventId) {
         Event event = findEventById(eventId);
@@ -67,8 +63,8 @@ public class EventService {
         return EventMapper.INSTANCE.toDtos(events);
     }
 
-    public Set<EventResponse> allForProject(Integer memberId) {
-        Set<Event> eventSet = eventRepository.findByProjectId(memberId);
+    public Set<EventResponse> allForProject(Integer projectId) {
+        Set<Event> eventSet = eventRepository.findByProjectId(projectId);
         return EventMapper.INSTANCE.toDtos(eventSet);
     }
 
@@ -81,24 +77,32 @@ public class EventService {
         return EventMapper.INSTANCE.toDto(findEventById(id));
     }
 
-    public EventResponse createPrivate(EventRequest request, HttpHeaders headers) {
+    public EventResponse create(EventRequest request, HttpHeaders headers) throws Exception {
         UserResponse user = userServiceClient.getUserFromHeaders(headers).getBody();
+        assert user != null;
+        if((!user.getRole().equals("ADMIN") && !user.getRole().equals("INFL")) && request.isPublic()) {
+            throw new Exception("You have no rights (^_−)");
+        }
+
         EventType eventType = eventTypeRepository.findById(request.getEventTypeId()).orElseThrow(
                 () -> new EntityNotFoundException("EventType with id: " + request.getEventTypeId() + " not found"));
-        Status status = statusRepository.findById(request.getStatusId()).orElseThrow(
-                () -> new EntityNotFoundException("Status with id: " + request.getStatusId() + " not found"));
-        Event event = new Event();
         ProjectResponse projectResponse = projectServiceClient.getProjectById(request.getProjectId()).getBody();
 
+        Event event = new Event();
         event.setName(request.getName());
         event.setLink(request.getLink());
         event.setEventType(eventType);
-        event.setPublic(false);
-        event.setStatus(status);
+        event.setPublic(true);
+        if(request.isPublic()) {
+            Status status = statusRepository.findByName(DEFAULT_STATUS).orElseThrow(
+                    () -> new EntityNotFoundException("Status with name: " + DEFAULT_STATUS + " not found"));
+            event.setStatus(status);
+        } else {
+            event.setStatus(null);
+        }
         assert projectResponse != null;
         event.setProjectId(request.getProjectId());
         event.setDescription(request.getDescription());
-        assert user != null;
         event.setCreatedBy(user.getId());
         event.setCreatedAt(OffsetDateTime.now());
         event.setStartDate(request.getStartDate());
@@ -150,35 +154,13 @@ public class EventService {
         return eventRepository.save(event);
     }
 
-    public EventResponse createPublic(EventRequest request, HttpHeaders headers) {
-        UserResponse user = userServiceClient.getUserFromHeaders(headers).getBody();
-        EventType eventType = eventTypeRepository.findById(request.getEventTypeId()).orElseThrow(
-                () -> new EntityNotFoundException("EventType with id: " + request.getEventTypeId() + " not found"));
-
-        ProjectResponse projectResponse = projectServiceClient.getProjectById(request.getProjectId()).getBody();
-
-        Event event = new Event();
-        event.setName(request.getName());
-        event.setLink(request.getLink());
-        event.setEventType(eventType);
-        event.setPublic(true);
-        event.setStatus(null);
-        assert projectResponse != null;
-        event.setProjectId(request.getProjectId());
-        event.setDescription(request.getDescription());
-        assert user != null;
-        event.setCreatedBy(user.getId());
-        event.setCreatedAt(OffsetDateTime.now());
-        event.setStartDate(request.getStartDate());
-        event.setEndDate(request.getEndDate());
-
-        eventRepository.save(event);
-        return EventMapper.INSTANCE.toDto(event);
-    }
-
-    public EventResponse update(Integer id, EventRequest request, HttpHeaders headers) {
+    public EventResponse update(Integer id, EventRequest request, HttpHeaders headers) throws Exception {
         Event event = findEventById(id);
         UserResponse user = userServiceClient.getUserFromHeaders(headers).getBody();
+        assert user != null;
+        if (!Objects.equals(event.getCreatedBy(), user.getId())) {
+            throw new Exception("You have no rights (^_−)");
+        }
 
         EventType eventType = eventTypeRepository.findById(request.getEventTypeId()).orElseThrow(
                 () -> new EntityNotFoundException("EventType with id: " + request.getEventTypeId() + " not found"));
@@ -188,7 +170,6 @@ public class EventService {
         event.setEventType(eventType);
         event.setStatus(status);
         event.setDescription(request.getDescription());
-        assert user != null;
         event.setUpdatedBy(user.getId());
         event.setUpdatedAt(OffsetDateTime.now());
         event.setStartDate(request.getStartDate());
@@ -198,7 +179,14 @@ public class EventService {
         return EventMapper.INSTANCE.toDto(event);
     }
 
-    public void delete(Integer id) {
+    public void delete(Integer id, HttpHeaders headers) throws Exception {
+        UserResponse user = userServiceClient.getUserFromHeaders(headers).getBody();
+        Event event = findEventById(id);
+
+        assert user != null;
+        if (!Objects.equals(event.getCreatedBy(), user.getId())) {
+            throw new Exception("You have no rights (^_−)");
+        }
         eventRepository.deleteById(id);
     }
 }
